Keep guard tests dependency-free
